pyproject.toml
[tool.poetry]
name = "kahani"
version = "0.1.0"
description = ""
authors = ["Sameer Segal <sameersegal@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
pandas = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
